picture-to-speech-clean/requirements.txt
opencv-python
numpy
pyttsx3
google-cloud-texttospeech>=2.16.0
